feat: implement `esm.mjs` task
