--- src/modules/cjsToEsm.cjs
@@ -1,39 +1,0 @@
-const { createServer: createServerHttp } = require('http');
-const { release, version } = require('os');
-const path = require('path');
-require('./files/c');
-
-const random = Math.random();
-
-let unknownObject;
-
-if (random > 0.5) {
-  unknownObject = require('./files/a.json');
-} else {
-  unknownObject = require('./files/b.json');
-}
-
-console.log(`Release ${release()}`);
-console.log(`Version ${version()}`);
-console.log(`Path segment separator is "${path.sep}"`);
-
-console.log(`Path to current file is ${__filename}`);
-console.log(`Path to current directory is ${__dirname}`);
-
-const myServer = createServerHttp((_, res) => {
-  res.end('Request accepted');
-});
-
-const PORT = 3000;
-
-console.log(unknownObject);
-
-myServer.listen(PORT, () => {
-  console.log(`Server is listening on port ${PORT}`);
-  console.log('To terminate it, use Ctrl+C combination');
-});
-
-module.exports = {
-  unknownObject,
-  myServer,
-};
